Falta validar atributos e recheio dos metodos editar

+++ Sistema/Adubo.java
@@ -4,7 +4,7 @@ public class Adubo {
     private String marcaAdubo;
     private String tipoAdubo;
     private String duracaoAdubo;
-    private int quantidadeAduboPorHectar;
+    private double quantidadeAduboPorHectar;
 
     public String getDuracaoAdubo() {
         return duracaoAdubo;
@@ -15,7 +15,7 @@ public class Adubo {
     public String getTipoAdubo() {
         return tipoAdubo;
     }
-    public int getQuantidadeAduboPorHectar() {
+    public double getQuantidadeAduboPorHectar() {
         return quantidadeAduboPorHectar;
     }
     public void setDuracaoAdubo(String duracaoAdubo) {
@@ -27,7 +27,7 @@ public class Adubo {
     public void setTipoAdubo(String tipoAdubo) {
         this.tipoAdubo = tipoAdubo;
     }
-    public void setQuantidadeAduboPorHectar(int quantidadeAduboPorHectar) {
+    public void setQuantidadeAduboPorHectar(double quantidadeAduboPorHectar) {
         this.quantidadeAduboPorHectar = quantidadeAduboPorHectar;
     }
